requirements.txt
mcp>=1.1.0
httpx>=0.28.0
python-dotenv>=1.0.1
pytest>=8.0.0
